control de acceso en Categorias

+++ src/categorias/categorias.controller.ts
@@ -6,7 +6,11 @@ import {
   Patch,
   Param,
   Delete,
+  Request,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
+import { AuthGuard } from 'src/auth/auth.guard';
 import { CategoriasService } from './categorias.service';
 import { CreateCategoriaDto } from './dto/create-categoria.dto';
 import { UpdateCategoriaDto } from './dto/update-categoria.dto';
@@ -15,14 +19,17 @@ import { UpdateCategoriaDto } from './dto/update-categoria.dto';
 export class CategoriasController {
   constructor(private readonly categoriasService: CategoriasService) {}
 
+  @UseGuards(AuthGuard)
   @Post()
-  create(@Body() createCategoriaDto: CreateCategoriaDto) {
+  create(@Body() createCategoriaDto: CreateCategoriaDto, @Request() req) {
+    createCategoriaDto.cliente_id = req.user.sub;
     return this.categoriasService.create(createCategoriaDto);
   }
 
+  @UseGuards(AuthGuard)
   @Get()
-  findAll() {
-    return this.categoriasService.findAll();
+  findAll(@Request() req) {
+    return this.categoriasService.findByClienteId(req.user.sub);
   }
 
   @Get('/eliminados')
@@ -30,21 +37,38 @@ export class CategoriasController {
     return this.categoriasService.findEliminados();
   }
 
+  @UseGuards(AuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.categoriasService.findOne(+id);
+  async findOne(@Param('id') id: string, @Request() req) {
+    if (await this.categoriasService.checkIsOwner(req.user.sub, +id)) {
+      return this.categoriasService.findOne(+id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
+  @UseGuards(AuthGuard)
   @Patch(':id')
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateCategoriaDto: UpdateCategoriaDto,
+    @Request() req,
   ) {
-    return this.categoriasService.update(+id, updateCategoriaDto);
+    if (await this.categoriasService.checkIsOwner(req.user.sub, +id)) {
+      updateCategoriaDto.cliente_id = req.user.sub;
+      return this.categoriasService.update(+id, updateCategoriaDto);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 
+  @UseGuards(AuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.categoriasService.remove(+id);
+  async remove(@Param('id') id: string, @Request() req) {
+    if (await this.categoriasService.checkIsOwner(req.user.sub, +id)) {
+      return this.categoriasService.remove(+id);
+    } else {
+      throw new UnauthorizedException();
+    }
   }
 }
